requirements.txt
click==5.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
djangorestframework==3.1.3
fakeredis==0.6.2
fudge==1.1.0
gunicorn==19.3.0
redis==2.10.5
requests==2.8.1
rq==0.5.6
static3==0.6.1
